+++ src/ipc/commands.rs
@@ -1,20 +1,39 @@
+use clap::ArgAction;
 use std::path::PathBuf;
 
-use clap::Subcommand;
+use clap::{Args, Subcommand};
 use serde::{Deserialize, Serialize};
 
 #[derive(Subcommand, Debug, Serialize, Deserialize)]
-#[serde(tag = "type", rename_all = "snake_case")]
+#[serde(tag = "command", rename_all = "snake_case")]
 pub enum Command {
-    /// Return "ok"
+    /// Pong
     Ping,
 
-    /// Open the GTK inspector
+    /// Open the GTK inspector.
     Inspect,
 
-    /// Reload the config
+    /// Reload the config.
     Reload,
 
+    /// Load an additional CSS stylesheet.
+    /// The sheet is automatically hot-reloaded.
+    LoadCss {
+        /// The path to the sheet.
+        path: PathBuf,
+    },
+
+    /// Get and set reactive Ironvar values.
+    #[command(subcommand)]
+    Var(IronvarCommand),
+
+    /// Interact with a specific bar.
+    Bar(BarCommand),
+}
+
+#[derive(Subcommand, Debug, Serialize, Deserialize)]
+#[serde(tag = "subcommand", rename_all = "snake_case")]
+pub enum IronvarCommand {
     /// Set an `ironvar` value.
     /// This creates it if it does not already exist, and updates it if it does.
     /// Any references to this variable are automatically and immediately updated.
@@ -34,49 +53,69 @@ pub enum Command {
 
     /// Gets the current value of all `ironvar`s.
     List,
+}
 
-    /// Load an additional CSS stylesheet.
-    /// The sheet is automatically hot-reloaded.
-    LoadCss {
-        /// The path to the sheet.
-        path: PathBuf,
-    },
+#[derive(Args, Debug, Serialize, Deserialize)]
+pub struct BarCommand {
+    /// The name of the bar.
+    pub name: String,
 
-    /// Set the visibility of the bar with the given name.
+    #[command(subcommand)]
+    #[serde(flatten)]
+    pub subcommand: BarCommandType,
+}
+
+#[derive(Subcommand, Debug, Serialize, Deserialize)]
+#[serde(tag = "subcommand", rename_all = "snake_case")]
+pub enum BarCommandType {
+    // == Visibility == \\
+    /// Force the bar to be shown, regardless of current visibility state.
+    Show,
+    /// Force the bar to be hidden, regardless of current visibility state.
+    Hide,
+    /// Set the bar's visibility state via an argument.
     SetVisible {
-        ///Bar name to target.
-        bar_name: String,
-        /// The visibility status.
-        #[arg(short, long)]
+        /// The new visibility state.
+        #[clap(
+            num_args(1),
+            require_equals(true),
+            action = ArgAction::Set,
+        )]
         visible: bool,
     },
+    /// Toggle the current visibility state between shown and hidden.
+    ToggleVisible,
+    /// Get the bar's visibility state.
+    GetVisible,
 
-    /// Get the visibility of the bar with the given name.
-    GetVisible {
-        /// Bar name to target.
-        bar_name: String,
+    // == Popup visibility == \\
+    /// Open a popup, regardless of current state.
+    /// If opening this popup, and a different popup on the same bar is already open, the other is closed.
+    ShowPopup {
+        /// The configured name of the widget.
+        widget_name: String,
     },
+    /// Close a popup, regardless of current state.
+    HidePopup,
+    /// Set the popup's visibility state via an argument.
+    /// If opening this popup, and a different popup on the same bar is already open, the other is closed.
+    SetPopupVisible {
+        /// The configured name of the widget.
+        widget_name: String,
 
+        #[clap(
+            num_args(1),
+            require_equals(true),
+            action = ArgAction::Set,
+        )]
+        visible: bool,
+    },
     /// Toggle a popup open/closed.
     /// If opening this popup, and a different popup on the same bar is already open, the other is closed.
     TogglePopup {
-        /// The name of the monitor the bar is located on.
-        bar_name: String,
-        /// The name of the widget.
-        name: String,
+        /// The configured name of the widget.
+        widget_name: String,
     },
-
-    /// Open a popup, regardless of current state.
-    OpenPopup {
-        /// The name of the monitor the bar is located on.
-        bar_name: String,
-        /// The name of the widget.
-        name: String,
-    },
-
-    /// Close a popup, regardless of current state.
-    ClosePopup {
-        /// The name of the monitor the bar is located on.
-        bar_name: String,
-    },
+    /// Get the popup's current visibility state.
+    GetPopupVisible,
 }
